move pokemon resources out of src
