+++ src/core/pkm/COLOPKM.ts
@@ -1,4 +1,5 @@
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
+import { Gen3ContestRibbons, Gen3StandardRibbons } from '@openhome-core/pokemon-resources'
 import { FourMoves } from '@openhome-core/util/types'
 import {
   AbilityNumber,
@@ -15,7 +16,6 @@ import {
   SpeciesLookup,
 } from '@pkm-rs/pkg'
 import { NationalDex } from '@pokemon-resources/consts/NationalDex'
-import { Gen3ContestRibbons, Gen3StandardRibbons } from '@pokemon-resources/index'
 import * as byteLogic from '../util/byteLogic'
 import * as stringLogic from '../util/stringConversion'
 import * as types from '../util/types'
