+++ src/core/pkm/PK4.ts
@@ -1,4 +1,5 @@
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
+import { Gen4Ribbons } from '@openhome-core/pokemon-resources'
 import { FourMoves } from '@openhome-core/util/types'
 import {
   AbilityIndex,
@@ -15,7 +16,6 @@ import {
   ShinyLeaves,
   SpeciesLookup,
 } from '@pkm-rs/pkg'
-import { Gen4Ribbons } from '@pokemon-resources/index'
 import * as byteLogic from '../util/byteLogic'
 import * as stringLogic from '../util/stringConversion'
 import * as types from '../util/types'
