+++ src/core/pkm/PK7.ts
@@ -1,5 +1,5 @@
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
-import { ModernRibbons } from '@openhome-core/pokemon-resources/src'
+import { ModernRibbons } from '@openhome-core/pokemon-resources/'
 import { FourMoves, PKMDate, Stats } from '@openhome-core/util/types'
 import {
   AbilityIndex,
