+++ src/core/pokemon-resources/consts/EncounterTypes.ts
