+++ src/core/pokemon-resources/consts/Forms.ts
