+++ src/core/pokemon-resources/consts/Items.ts
