+++ src/core/pokemon-resources/consts/NationalDex.ts
