+++ src/core/pokemon-resources/consts/Ribbons.ts
@@ -1,4 +1,4 @@
-import { Gen34ContestRibbons, Gen34TowerRibbons } from '@pokemon-resources/index'
+import { Gen34ContestRibbons, Gen34TowerRibbons } from '@openhome-core/pokemon-resources'
 
 export const Gen9Ribbons = [
   'Kalos Champion',
