+++ src/core/pokemon-resources/consts/ShadowIDs.ts
