+++ src/core/pokemon-resources/consts/TransferRestrictions.ts
