+++ src/core/pokemon-resources/index.ts
