+++ src/core/pokemon-resources/learned-moves/bdsp-tms.ts
