+++ src/core/pokemon-resources/learned-moves/sv-tms.ts
