+++ src/core/pokemon-resources/learned-moves/swsh-trs.ts
