+++ src/core/pokemon-resources/memories/gen6/feelings.ts
