+++ src/core/pokemon-resources/memories/gen6/locations.ts
