+++ src/core/pokemon-resources/other/characteristics.ts
