+++ src/core/pokemon-resources/other/index.ts
