+++ src/core/pokemon-resources/other/ribbons.ts
