+++ src/core/pokemon-resources/other/types.ts
