+++ src/ui/pokemon-details/tabs/MetDataMovesTab.tsx
@@ -2,6 +2,7 @@ import { PKMInterface } from '@openhome-core/pkm/interfaces'
 import { getLocationString, getLocationStringOrOrigin } from '@openhome-core/pkm/MetLocation'
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
 import { getCharacteristic, getMoveMaxPP } from '@openhome-core/pkm/util'
+import { RibbonTitles } from '@openhome-core/pokemon-resources'
 import { getPluginIdentifier } from '@openhome-core/save/util'
 import Markings from '@openhome-ui/components/pokemon/Markings'
 import MoveCard from '@openhome-ui/components/pokemon/MoveCard'
@@ -9,7 +10,6 @@ import { getPublicImageURL } from '@openhome-ui/images/images'
 import { getBallIconPath } from '@openhome-ui/images/items'
 import { AppInfoContext } from '@openhome-ui/state/appInfo'
 import { Language, Languages, OriginGames } from '@pkm-rs/pkg'
-import { RibbonTitles } from '@pokemon-resources/index'
 import { Badge, Flex, Grid } from '@radix-ui/themes'
 import { useContext, useMemo } from 'react'
 import './MetDataMovesTab.css'
